Making a uploadForm and progressBar as well as a hook for communicating with the firebase.storage.
We flipflopped between using
stateless component and classbased
components for upload and wound
up with another error.

The stateless approach was good,
more simple and easily managed,
but we changed to class based
because we were having trouble
accessing inputted file with
the original scheme. I think
stateless is the way but we will need to
find a better way to read input file.