Changed project name in package.json to what its supposed to be 'oral-history'
Added some technical info to README.md as well as about info
for the layperson.

Added documentation to useFirestore hook.

Finished our image grid, which displays
the images stored in storage and firestore.

Added a 'Modal' component which
displays a single image when a user
clicks on it, and disappears when the
user clicks away from it.

Realizing we need a description as
well as the url for our image, Added
a text area and some logic for controlling
that change event, its commented  out
until we finish everything else and are
moving onto personalizing/ customizing
the site.

Things are comming together here, one
more tutorial that will introduce animations, then we can move on
to adding routing, authentication,
user permissions, commenting, etc.