Added quite a bit of routing and styling:
nav bar and routing working, in their
own component title,

note, css problem with unexpected import
this could be because you import
title in app.js, which imports browse
photos and so photos.css. Not sure
how to prevent this style bubbling but
it is quite undesirable and difficult to mitigate.

Added a directory structure, component
folder is now split into general, auth
and photos.

made a signup form component
unable to test it because of the above
mentioned css import issue, which
makes these inputs invisible unless
i manually switch the style off in
browser dev tools. want to solve
that import problem before moving
on, but the signup form should be
complete with validation and everything
just need to get it set up with a hook
that creates an auth token and saves
the user to the database profiles collection.