Ditched the latest design moodswing.
Navbar, and sign out sign in functionality
is in its own component, which
is then a child of auth provider within
app.js. Site is working, ran into a
uncaught bug while testing, that makes
little sense, but glad to see we could
figure out what the previous bugs cause is

Going to be taking a 4-5 day break from
coding while I go on vacation. Back at it
on the regular and commiting heavy (finishing this repo asap) soon
as i get back on saturday...